Documentation for base styles
