+++ scss/bitstyles/base/_all.scss
@@ -1,7 +1,7 @@
-@import 'typography';
-@import 'link';
-@import 'images';
-@import 'forms';
-@import 'media';    // FOR TESTING ONLY - Empty of styles
-@import 'hr';       // FOR TESTING ONLY - Empty of styles
-@import 'figure';   // FOR TESTING ONLY - Empty of styles
+@import 'typography/';
+@import 'link/';
+@import 'images/';
+@import 'forms/';
+@import 'media/';    // FOR TESTING ONLY - Empty of styles
+@import 'hr/';       // FOR TESTING ONLY - Empty of styles
+@import 'figure/';   // FOR TESTING ONLY - Empty of styles
