--- scss/bitstyles/base/_figure.scss
@@ -1,9 +1,0 @@
-// Figure and figcaption
-//
-// markup:
-// <figure>
-//   <img src="assets/images/placeholder_16-9.png">
-//   <figcaption>An image at 16:9 aspect ratio</figcaption>
-// </figure>
-//
-// Styleguide Base.figure
