--- scss/bitstyles/base/_forms.scss
@@ -1,187 +1,0 @@
-/* stylelint-disable selector-max-type */
-// Forms
-//
-
-// fieldset
-//
-// markup:
-// <fieldset><legend>Fieldset</legend></fieldset>
-//
-// Styleguide Base.fieldset
-
-// input - text
-//
-// markup:
-// <input type="text" placeholder="Jelly cake marshmallow">
-//
-// Styleguide Base.input.text
-
-// input - email
-//
-// markup:
-// <input type="email" placeholder="jellycake@marshmallow.com">
-//
-// Styleguide Base.input.email
-
-// input - number
-//
-// markup:
-// <input type="number" placeholder="123456789">
-//
-// Styleguide Base.input.number
-
-// input - range
-//
-// markup:
-// <input type="range" min="0" max="10" step="1">
-//
-// Styleguide Base.input.range
-
-// textarea
-//
-// markup:
-// <textarea>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</textarea>
-//
-// Styleguide Base.textarea
-
-textarea {
-  resize: vertical;
-}
-
-// select
-//
-// markup:
-// <select>
-//   <option value="value1" selected>Value 1</option>
-//   <option value="value2">Value 2</option>
-//   <option value="value3">Value 3</option>
-// </select>
-//
-// Styleguide Base.select
-
-// label
-//
-// markup:
-// <label>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</label>
-//
-// Styleguide Base.label
-
-// input - radio
-//
-// markup:
-// <input type="radio">
-//
-// Styleguide Base.input.radio
-
-// input - checkbox
-//
-// markup:
-// <input type="checkbox">
-//
-// Styleguide Base.input.checkbox
-
-// input - button
-//
-// markup:
-// <input type="button" value="Button">
-//
-// Styleguide Base.input.button
-
-// input - color
-//
-// markup:
-// <input type="color">
-//
-// Styleguide Base.input.color
-
-// input - date
-//
-// markup:
-// <input type="date">
-//
-// Styleguide Base.input.date
-
-// input - datetime-local
-//
-// markup:
-// <input type="datetime-local">
-//
-// Styleguide Base.input.datetime-local
-
-// input - file
-//
-// markup:
-// <input type="file">
-//
-// Styleguide Base.input.file
-
-// input - image
-//
-// markup:
-// <input type="image" src="assets/images/placeholder_button.png" alt="Image">
-//
-// Styleguide Base.input.image
-
-// input - month
-//
-// markup:
-// <input type="month">
-//
-// Styleguide Base.input.month
-
-// input - password
-//
-// markup:
-// <input type="password" value="Password">
-//
-// Styleguide Base.input.password
-
-// input - reset
-//
-// markup:
-// <input type="reset">
-//
-// Styleguide Base.input.reset
-
-// input - search
-//
-// markup:
-// <input type="search">
-//
-// Styleguide Base.input.search
-
-// input - submit
-//
-// markup:
-// <input type="submit" value="Submit">
-//
-// Styleguide Base.input.submit
-
-// input - tel
-//
-// markup:
-// <input type="tel">
-//
-// Styleguide Base.input.tel
-
-// input - time
-//
-// markup:
-// <input type="time">
-//
-// Styleguide Base.input.type
-
-// input - url
-//
-// markup:
-// <input type="url">
-//
-// Styleguide Base.input.url
-
-// input - week
-//
-// markup:
-// <input type="week">
-//
-// Styleguide Base.input.week
-/* stylelint-enable selector-max-type */
