--- scss/bitstyles/base/_hr.scss
@@ -1,6 +1,0 @@
-// Horizontal rule
-//
-// markup:
-// <hr>
-//
-// Styleguide Base.hr
