--- scss/bitstyles/base/_link.scss
@@ -1,20 +1,0 @@
-/* stylelint-disable selector-max-type */
-// Links
-//
-// Standard links.
-//
-// markup:
-// <a href="#"" class="{$modifiers}">Jelly cake marshmallow</a>
-//
-// default        - Default state
-// :hover         - Hover state
-// :focus         - Focus state
-// :visited       - Visited state
-//
-// Styleguide Base.link
-
-@import '../settings/link';
-
-a {
-  @include link;
-}
