--- scss/bitstyles/base/_media.scss
@@ -1,24 +1,0 @@
-// Audio
-//
-// markup:
-// <audio controls="">
-//   <source src="http://hpr.dogphilosophy.net/test/opus.opus" type="audio/ogg; codecs=&quot;opus&quot;">
-//   <source src="http://hpr.dogphilosophy.net/test/weba.weba">
-//   <source src="http://hpr.dogphilosophy.net/test/ogg.ogg">
-//   <source src="http://hpr.dogphilosophy.net/test/mp3.mp3">
-//   <source src="http://hpr.dogphilosophy.net/test/wav.wav">
-//   <source src="http://hpr.dogphilosophy.net/test/flac.flac">
-// </audio>
-//
-// Styleguide Base.audio
-
-// Video
-//
-// markup:
-// <video width="640" height="360" controls="true" >
-// 	<source src="http://clips.vorwaerts-gmbh.de/VfE_html5.mp4" type="video/mp4">
-// 	<source src="http://clips.vorwaerts-gmbh.de/VfE.webm" type="video/webm">
-// 	<source src="http://clips.vorwaerts-gmbh.de/VfE.ogv" type="video/ogg">
-// </video>
-//
-// Styleguide Base.video
