--- scss/bitstyles/base/_typography.scss
@@ -1,296 +1,0 @@
-/* stylelint-disable selector-max-type */
-html {
-  font-family: $bitstyles-font-family-body;
-  line-height: $bitstyles-line-height-base;
-  color: palette('text');
-  background: palette('background');
-
-  &.fonts-loaded {
-    font-family: $bitstyles-font-family-body-loaded;
-  }
-}
-
-// Headings
-//
-// Semantic headings, `<h1>` being the most important in a section or page, `<h6>` being the least.
-//
-// markup:
-// <h1>Header 1</h1>
-// <h2>Header 2</h2>
-// <h3>Header 3</h3>
-// <h4>Header 4</h4>
-// <h5>Header 5</h5>
-// <h6>Header 6</h6>
-//
-// Styleguide Base.headings
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  @include generic-heading;
-}
-
-@include generate-font-sizes('', (html, h1, h2, h3, h4, h5, h6));
-
-// Paragraphs
-//
-// A paragraph of text.
-//
-// markup:
-// <p>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer. Gummies cheesecake oat cake sugar plum icing cupcake tiramisu bonbon. Cotton candy chupa chups tootsie roll chupa chups cotton candy liquorice jelly gingerbread. Pastry gummi bears liquorice tart cotton candy marshmallow. Ice cream cotton candy chocolate cake cookie. Bonbon candy jelly-o sugar plum cotton candy carrot cake icing ice cream. Sweet chocolate marzipan. Candy canes danish cake carrot cake bonbon. Gummi bears sesame snaps tart bear claw pie chocolate bar. Ice cream candy canes sugar plum cookie. Macaroon biscuit biscuit carrot cake liquorice. Muffin pudding gingerbread powder jelly-o chocolate bar powder jelly beans toffee.</p>
-//
-// Styleguide Base.paragraph
-
-p {
-  margin: 0 0 spacing('base');
-}
-
-// Address
-//
-// Address element. Often used to indicate authorship and contact information.
-//
-// markup:
-// <address>Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</address
-//
-// Styleguide Base.address
-
-address {
-  font-style: normal;
-}
-
-// Definition list
-//
-// Used to define terms.
-//
-// markup:
-// <dl>
-//   <dt>Gummies cheesecake oat cake sugar plum icing cupcake tiramisu bonbon.</dt>
-//   <dd>Gummies cheesecake oat cake sugar plum icing cupcake tiramisu bonbon.</dd>
-// </dl>
-//
-// Styleguide Base.dl
-
-dt {
-  font-weight: $bitstyles-font-weight-bold;
-}
-
-// List - Unordered
-//
-// An unordered list.
-//
-// markup:
-// <ul>
-//   <li>Pie cotton candy chupa chups.</li>
-//   <li>Liquorice cheesecake wafer.</li>
-//   <li>Gummies cheesecake oat cake sugar.</li>
-// </ul>
-//
-// Styleguide Base.ul
-
-// List - Ordered
-//
-// An unordered list.
-//
-// markup:
-// <ol>
-//   <li>Pie cotton candy chupa chups.</li>
-//   <li>Liquorice cheesecake wafer.</li>
-//   <li>Gummies cheesecake oat cake sugar.</li>
-// </ol>
-//
-// Styleguide Base.ol
-
-// Blockquote
-//
-// Quoted text.
-//
-// markup:
-// <blockquote>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer. Gummies cheesecake oat cake sugar plum icing cupcake tiramisu bonbon. Cotton candy chupa chups tootsie roll chupa chups cotton candy liquorice jelly</blockquote>
-//
-// Styleguide Base.blockquote
-
-// strong
-//
-// markup:
-// <strong>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer. Gummies cheesecake oat cake sugar plum icing cupcake tiramisu bonbon. Cotton candy chupa chups tootsie roll chupa chups cotton candy liquorice jelly</strong>
-//
-// Styleguide Base.strong
-
-// em
-//
-// markup:
-// <em>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</em>
-//
-// Styleguide Base.em
-
-// pre
-//
-// markup:
-// <pre>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</pre>
-//
-// Styleguide Base.pre
-
-// code
-//
-// markup:
-// <code>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</code>
-//
-// Styleguide Base.code
-
-// kbd
-//
-// markup:
-// <kbd>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</kbd>
-//
-// Styleguide Base.markup
-
-// samp
-//
-// markup:
-// <samp>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</samp>
-//
-// Styleguide Base.samp
-
-// var
-//
-// markup:
-// <var>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</var>
-//
-// Styleguide Base.var
-
-// cite
-//
-// markup:
-// <cite>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</cite>
-//
-// Styleguide Base.cite
-
-// del
-//
-// markup:
-// <del>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</del>
-//
-// Styleguide Base.del
-
-// abbr
-//
-// markup:
-// <abbr>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</abbr>
-//
-// Styleguide Base.abbr
-
-// i
-//
-// markup:
-// <i>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</i>
-//
-// Styleguide Base.i
-
-// b
-//
-// markup:
-// <b>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</b>
-//
-// Styleguide Base.b
-
-// ins
-//
-// markup:
-// <ins>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</ins>
-//
-// Styleguide Base.ins
-
-// dfn
-//
-// markup:
-// <dfn>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</dfn>
-//
-// Styleguide Base.dfn
-
-// small
-//
-// markup:
-// <small>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</small>
-//
-// Styleguide Base.small
-
-// sub
-//
-// markup:
-// <sub>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</sub>
-//
-// Styleguide Base.sub
-
-// sup
-//
-// markup:
-// <sup>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</sup>
-//
-// Styleguide Base.sup
-
-// big
-//
-// markup:
-// <big>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</big>
-//
-// Styleguide Base.big
-
-// s
-//
-// markup:
-// <s>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</s>
-//
-// Styleguide Base.s
-
-// strike
-//
-// markup:
-// <strike>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</strike>
-//
-// Styleguide Base.strike
-
-// q
-//
-// markup:
-// <q>Jelly cake marshmallow. Pie cotton candy chupa chups marzipan liquorice cheesecake wafer.</q>
-//
-// Styleguide Base.q
-
-// time
-//
-// markup:
-// <time>20:57</time>
-//
-// Styleguide Base.time
-
-// Table
-//
-// markup:
-// <table>
-//   <caption>Awesome caption</caption>
-//   <thead>
-//     <tr>
-//       <th>Header content 1</th>
-//       <th>Header content 2</th>
-//     </tr>
-//   </thead>
-//   <tfoot>
-//     <tr>
-//       <td>Footer content 1</td>
-//       <td>Footer content 2</td>
-//     </tr>
-//   </tfoot>
-//   <tbody>
-//     <tr>
-//       <td>Body content 1</td>
-//       <td>Body content 2</td>
-//     </tr>
-//   </tbody>
-// </table>
-//
-// Styleguide Base.table
-
-/* stylelint-enable selector-max-type */
