+++ scss/bitstyles/base/figure/figure.md
@@ -1,0 +1,1 @@
+# Figure and figcaption
