+++ scss/bitstyles/base/forms/_.scss
@@ -1,0 +1,5 @@
+/* stylelint-disable selector-max-type */
+textarea {
+  resize: vertical;
+}
+/* stylelint-enable selector-max-type */
