+++ scss/bitstyles/base/forms/forms.md
@@ -1,0 +1,52 @@
+# Forms
+
+## fieldset
+
+
+## input - text
+
+## input - email
+
+## input - number
+
+## input - range
+
+## textarea
+
+## select
+
+## label
+
+## input - radio
+
+## input - checkbox
+
+## input - button
+
+## input - color
+
+## input - date
+
+## input - datetime-local
+
+## input - file
+
+## input - image
+
+## input - month
+
+## input - password
+
+## input - reset
+
+## input - search
+
+## input - submit
+
+## input - tel
+
+## input - time
+
+## input - url
+
+## input - week
