+++ scss/bitstyles/base/hr/hr.md
@@ -1,0 +1,1 @@
+# Horizontal rule
