+++ scss/bitstyles/base/hr/hr.stories.js
@@ -1,0 +1,14 @@
+import Readme from './hr.md';
+
+export default {
+  title: 'Base/Hr',
+  parameters: {
+    readme: {
+      sidebar: Readme,
+    },
+  },
+};
+
+export const hr = () => `
+  <hr>
+`;
