+++ scss/bitstyles/base/images/_.scss
@@ -1,10 +1,3 @@
-// Image
-//
-// markup:
-// <img src="assets/images/placeholder_16-9.png">
-//
-// Styleguide Base.img
-
 /* stylelint-disable selector-max-type */
 img {
   height: auto;
