+++ scss/bitstyles/base/images/images.md
@@ -1,0 +1,3 @@
+# Image
+
+Does the minimum to make images responsive — sets `max-width: 100%`
