+++ scss/bitstyles/base/link/_.scss
@@ -1,0 +1,7 @@
+@import '../../settings/link';
+
+/* stylelint-disable selector-max-type */
+a {
+  @include link;
+}
+/* stylelint-enable selector-max-type */
