+++ scss/bitstyles/base/link/link.md
@@ -1,0 +1,3 @@
+# Links
+
+Standard links.
