+++ scss/bitstyles/base/link/link.stories.js
@@ -1,0 +1,14 @@
+import Readme from './link.md';
+
+export default {
+  title: 'Base/Link',
+  parameters: {
+    readme: {
+      sidebar: Readme,
+    },
+  },
+};
+
+export const link = () => `
+  <a href="#"">Jelly cake marshmallow</a>
+`;
