+++ scss/bitstyles/base/media/_.scss
@@ -1,0 +1,1 @@
+/* stylelint-disable no-empty-source */
