+++ scss/bitstyles/base/typography/typography.md
@@ -1,0 +1,33 @@
+# Typography
+
+## Headings
+
+Semantic headings, `<h1>` being the most important in a section or page, `<h6>` being the least.
+
+## Paragraphs
+
+A paragraph of text.
+
+## Address
+
+Address element. Often used to indicate authorship and contact information.
+
+// Definition list
+//
+// Used to define terms.
+//
+
+// List - Unordered
+//
+// An unordered list.
+//
+
+// List - Ordered
+//
+// An ordered list.
+//
+
+// Blockquote
+//
+// Quoted text, ideally with citation
+//
