--- test/stories/base/figure.stories.js
@@ -1,8 +1,0 @@
-export default { title: 'Base/Figure' };
-
-export const figure = () => `
-  <figure>
-   <img src="https://placekitten.com/200/300">
-   <figcaption>An portrait of a kitten</figcaption>
-  </figure>
-`;
