--- test/stories/base/hr.stories.js
@@ -1,5 +1,0 @@
-export default { title: 'Base/Hr' };
-
-export const hr = () => `
-  <hr>
-`;
