--- test/stories/base/images.stories.js
@@ -1,5 +1,0 @@
-export default { title: 'Base/Images' };
-
-export const img = () => `
-  <img src="assets/images/placeholder_16-9.png"></img>
-`;
