--- test/stories/base/link.stories.js
@@ -1,5 +1,0 @@
-export default { title: 'Base/Link' };
-
-export const link = () => `
-  <a href="#"">Jelly cake marshmallow</a>
-`;
